+++ examples/python_gtmetrix2.py
@@ -1,1 +1,0 @@
-../src/python_gtmetrix2/__init__.py